Fix finding Glslang to find SPIRV
SPIRV is what's needed in the source files, but it doesn't have a FindXXX module, or a cmake file at all.
This looks for it on the system, and can also take a user-inputted SPIRV_ROOT param to help find it.

It is still optional, and the in-tree Glslang build will still work if it wasn't found.