Fixed an issue with joystick initalization.
Fixed an issue where having a joystick connected during startup would
cause an exception.

This was caused by the joystick module constructor using the callback
which would refer back to the module. As the module was still in the
process of being initialized, this caused the sanitycheck to fail
and the engine to abort.

Because this issue was introduced while cleaning up this code,
a comment has been left to prevent this issue from rising again.