Fixed a linker issue with SPIRV/GLSLang.
Fixed an issue under Unix environments which caused the
final applications to throw a linker error when linking with libAcid.

This only occurred when an installation of libSPIRV/libGLSLang was found
on the system. A difference between using CMake for external projects
(where transitive dependencies are automatically taken into account) and
manually linking against static libraries (e.g. system-wide
installations of GLSLang under Ubuntu) was causing the test applications
in the latter case to still have unresolved symbols.

Eventually, it might be wise to enable checking for unresolved symbols
in libAcid.so, but for now this will do.

Furthermore, silenced some warnings for missing packages (which were
then downloaded anyway).