Made modules self registering and added requires to Register for initialization order.
Input requires Joysticks, Keyboard, and Mouse, each of those modules requires Window; Window will be created before those other modules.

I'm not sure if Module requires are just going to be for initialization, or will also reflect Update requirements (update scene before shadows etc.). Because a filter is allowed for the Engine some Module dependencies can be broken by the application.

Module order in a Stage update may also shift depending on the compiler, if a Module does not require another but __depends__ on it.