Fix destruction order issues
A lot of things call back into half-destructed objects, leading to crashes.
For example, the CommandPool destructor has

	auto logicalDevice = Graphics::Get()->GetLogicalDevice();

but that is null while destroying the commandPools member of Graphics.
Moving the destruction a little earlier fixes the problem.

The same issue exists for the tutorial apps, which have their destructors
reference parts of Engine which is already in the process of being
destroyed.