Allow top level input scheme signals to be moved when switching.
Can't dynamic cast the result of `InputScheme::GetAxis` or `InputScheme::GetButton` and try to set implementation signals manually as on scheme switching those callbacks will be lost.
If a new scheme does not implement a mapped name the signal will not be moved, so the downstream slots will be lost.